Worker moved to new worker.py
